Luke Gorrie: Add support for QEMU vhost-user feature
This patch adds support for the QEMU vhost-user feature to Libvirt.
vhost-user is an efficient interface between guests and userspace
network stacks (vswitches).

Our goal is to have vhost-user supported in the OpenStack Juno
release. We hope it is okay to start code review on qemu, Libvirt, and
OpenStack in parallel. We understand that code will have to be merged
first into QEMU and then into Libvirt and then into OpenStack.

References:

  vhost-user qemu patches:
    http://lists.gnu.org/archive/html/qemu-devel/2014-05/msg05443.html
  vhost-user high level description:
    http://www.virtualopensystems.com/en/solutions/guides/snabbswitch-qemu/
  OpenStack feature that this is intended to support (Snabb NFV):
    http://snabb.co/nfv.html
    https://blueprints.launchpad.net/neutron/+spec/snabb-nfv-mech-driver

Signed-off-by: Michele Paolino <m.paolino@virtualopensystems.com>